[DF-11148] Code format, warning fix
Former-commit-id: 5a9d3a8dd19fd5344deecb41447cca0e3d0f13a1

+++ Sources/Internal/Render/RHI/DX9/rhi_CommandBufferDX9.cpp
@@ -170,7 +170,7 @@ struct CommandBufferDX9_t
     template <class... Args>
     void Command(Args... args)
     {
-        uint64_t argumentsCount = sizeof...(args);
+        size_t argumentsCount = sizeof...(args);
         _cmd.reserve(_cmd.size() + argumentsCount);
         _cmd.insert(_cmd.end(), { static_cast<uint64>(args)... });
     }
